feat:#87 整合node-siyuan到挂件-完善UI
