+++ src/components/blog/themes/default/DefaultHeader.vue
@@ -1,78 +1,10 @@
 <template>
   <div>
-    <el-autocomplete
-        v-model="state"
-        :fetch-suggestions="querySearch"
-        popper-class="my-autocomplete"
-        placeholder="Please input"
-        @select="handleSelect"
-    >
-      <template #suffix>
-        <el-icon class="el-input__icon" @click="handleIconClick">
-          <edit/>
-        </el-icon>
-      </template>
-      <template #default="{ item }">
-        <div class="value">{{ item.value }}</div>
-        <span class="link">{{ item.link }}</span>
-      </template>
-    </el-autocomplete>
-    <el-button class="s-btn" type="primary" @click="handleBtnSearch">搜索</el-button>
+    <h1 style="text-align: center;margin-top: 0;">思源笔记发布工具</h1>
   </div>
 </template>
 
 <script lang="ts" setup>
-import {onMounted, ref} from 'vue'
-import {Edit} from '@element-plus/icons-vue'
-
-interface LinkItem {
-  value: string
-  link: string
-}
-
-const state = ref('')
-const links = ref<LinkItem[]>([])
-
-const querySearch = (queryString: string, cb: any) => {
-  const results = queryString
-      ? links.value.filter(createFilter(queryString))
-      : links.value
-  // call callback function to return suggestion objects
-  cb(results)
-}
-const createFilter = (queryString: any) => {
-  return (restaurant: any) => {
-    return (
-        restaurant.value.toLowerCase().indexOf(queryString.toLowerCase()) === 0
-    )
-  }
-}
-const loadAll = () => {
-  return [
-    {value: 'vue', link: 'https://github.com/vuejs/vue'},
-    {value: 'element', link: 'https://github.com/ElemeFE/element'},
-    {value: 'cooking', link: 'https://github.com/ElemeFE/cooking'},
-    {value: 'mint-ui', link: 'https://github.com/ElemeFE/mint-ui'},
-    {value: 'vuex', link: 'https://github.com/vuejs/vuex'},
-    {value: 'vue-router', link: 'https://github.com/vuejs/vue-router'},
-    {value: 'babel', link: 'https://github.com/babel/babel'},
-  ]
-}
-const handleSelect = (item: LinkItem) => {
-  console.log(item)
-}
-
-const handleIconClick = (ev: Event) => {
-  console.log(ev)
-}
-
-const handleBtnSearch = () => {
-  console.log("handleBtnSearch")
-}
-
-onMounted(() => {
-  links.value = loadAll()
-})
 </script>
 
 <script lang="ts">
@@ -82,26 +14,4 @@ export default {
 </script>
 
 <style scoped>
-.my-autocomplete li {
-  line-height: normal;
-  padding: 7px;
-}
-
-.my-autocomplete li .name {
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-
-.my-autocomplete li .addr {
-  font-size: 12px;
-  color: #b4b4b4;
-}
-
-.my-autocomplete li .highlighted .addr {
-  color: #ddd;
-}
-
-.s-btn{
-  margin-left: 20px;
-}
 </style>
